pytest revamp setting up removal of local models
